Restructure: extract time, ids, endpoints, label into own modules
Per the rust-project-structure conventions (P-DIR-MOD, P-API): lib.rs
should mostly re-export from child modules, not host all definitions
inline. First batch — pure leaf types with no algorithm dependencies:

- time.rs: SecondOfDay, Duration, Transfers (newtypes + arithmetic impls)
- ids.rs: StopIdx, RouteIdx, TripIdx (dense u32 indices)
- endpoints.rs: Endpoints + IntoEndpoints (origin/target input shapes)
- label.rs: Label trait + ArrivalTime (single-criterion default impl)

lib.rs gains the mod declarations and pub use re-exports. External
API surface is identical (raptor::SecondOfDay, raptor::Label etc. all
still resolve at the crate root). 64 unit tests + 18 proptests still
green.
