+++ raptor/src/lib.rs
@@ -138,12 +138,17 @@
 //! - `internal` — enables the `raptor` criterion benchmark over [`manual`].
 
 use std::cmp::Reverse;
-use std::collections::{BTreeMap, BinaryHeap};
-use std::fmt;
+use std::collections::BTreeMap;
+use std::collections::BinaryHeap;
 use std::sync::Mutex;
 
 use fixedbitset::FixedBitSet;
 use smallvec::SmallVec;
+
+mod endpoints;
+mod ids;
+mod label;
+mod time;
 
 pub mod gtfs;
 pub mod labels;
@@ -152,561 +157,23 @@ pub mod labels;
 /// GTFS feed — for tests, custom data sources, and toy examples.
 pub mod manual;
 
+pub use endpoints::Endpoints;
+pub use endpoints::IntoEndpoints;
+pub use ids::RouteIdx;
+pub use ids::StopIdx;
+pub use ids::TripIdx;
+pub use label::ArrivalTime;
+pub use label::Label;
+pub use time::Duration;
+pub use time::SecondOfDay;
+pub use time::Transfers;
+
 #[cfg(test)]
 mod test;
 
 /// Internal round-counter type, used only for indexing label arrays.
 /// User-facing transfer caps are passed as [`Transfers`].
 pub(crate) type K = usize;
-
-/// A point in time, in seconds since midnight on the timetable's
-/// service date. Wraps a `u32` — the day is 86,400 seconds; `u32`
-/// covers feed quirks like trips encoded past 24h with room to spare.
-///
-/// SecondOfDay is a *timestamp*. A *length* of time — walk-time offset,
-/// transfer time, dwell time — is a [`Duration`], a distinct type.
-/// The trait surface uses both consistently so they can't be
-/// silently confused.
-///
-/// Construct via [`SecondOfDay::ZERO`], [`SecondOfDay::from_secs`], [`SecondOfDay::hms`], or
-/// the public-field constructor `SecondOfDay(n)`. Extract via
-/// [`SecondOfDay::as_secs`] / [`SecondOfDay::as_hms`]. Arithmetic with [`Duration`]
-/// is saturating: `SecondOfDay::MAX + Duration::MAX` stays at `SecondOfDay::MAX`.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
-pub struct SecondOfDay(pub u32);
-
-impl SecondOfDay {
-    /// Midnight (0 seconds since the start of the service day).
-    pub const ZERO: SecondOfDay = SecondOfDay(0);
-    /// Sentinel for "unreached". The algorithm uses this internally
-    /// for empty `(round, stop)` cells.
-    pub const MAX: SecondOfDay = SecondOfDay(u32::MAX);
-
-    /// Construct from a raw seconds-since-midnight count.
-    pub const fn from_secs(s: u32) -> Self {
-        SecondOfDay(s)
-    }
-
-    /// Construct from `(hours, minutes, seconds)`.
-    pub const fn hms(h: u32, m: u32, s: u32) -> Self {
-        SecondOfDay(h * 3600 + m * 60 + s)
-    }
-
-    /// The underlying `u32` — seconds since midnight.
-    pub const fn as_secs(self) -> u32 {
-        self.0
-    }
-
-    /// `(hours, minutes, seconds)` decomposition.
-    pub const fn as_hms(self) -> (u32, u32, u32) {
-        (self.0 / 3600, (self.0 / 60) % 60, self.0 % 60)
-    }
-
-    /// Iterator of departure times from `start` (inclusive) to `end`
-    /// (exclusive) at `step`-second intervals. Convenience for the
-    /// common range-query input pattern:
-    ///
-    /// ```
-    /// # use raptor::SecondOfDay;
-    /// let deps: Vec<_> = SecondOfDay::every(
-    ///     SecondOfDay::hms(17, 0, 0),
-    ///     SecondOfDay::hms(17, 5, 0),
-    ///     60,
-    /// ).collect();
-    /// assert_eq!(deps.len(), 5);
-    /// assert_eq!(deps[0], SecondOfDay::hms(17, 0, 0));
-    /// assert_eq!(deps[4], SecondOfDay::hms(17, 4, 0));
-    /// ```
-    ///
-    /// Pass directly to
-    /// [`Query::depart_in_window`](crate::Query::depart_in_window).
-    pub fn every(
-        start: SecondOfDay,
-        end: SecondOfDay,
-        step: u32,
-    ) -> impl Iterator<Item = SecondOfDay> + Clone {
-        (start.0..end.0)
-            .step_by(step.max(1) as usize)
-            .map(SecondOfDay)
-    }
-}
-
-impl From<u32> for SecondOfDay {
-    fn from(s: u32) -> Self {
-        SecondOfDay(s)
-    }
-}
-
-impl fmt::Display for SecondOfDay {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl std::ops::Add<Duration> for SecondOfDay {
-    /// `SecondOfDay + Duration` advances the timestamp; saturating on overflow.
-    type Output = SecondOfDay;
-    fn add(self, d: Duration) -> SecondOfDay {
-        SecondOfDay(self.0.saturating_add(d.0))
-    }
-}
-
-impl std::ops::Sub<SecondOfDay> for SecondOfDay {
-    /// `SecondOfDay - SecondOfDay` is the [`Duration`] between them, saturating at
-    /// [`Duration::ZERO`] if `self < other`.
-    type Output = Duration;
-    fn sub(self, other: SecondOfDay) -> Duration {
-        Duration(self.0.saturating_sub(other.0))
-    }
-}
-
-impl std::ops::Sub<Duration> for SecondOfDay {
-    /// `SecondOfDay - Duration` rewinds the timestamp; saturating at
-    /// [`SecondOfDay::ZERO`] on underflow.
-    type Output = SecondOfDay;
-    fn sub(self, d: Duration) -> SecondOfDay {
-        SecondOfDay(self.0.saturating_sub(d.0))
-    }
-}
-
-/// A length of time, in seconds. Distinct from [`SecondOfDay`] (a point in
-/// time) so the algorithm signatures can express which kind they
-/// expect: a walk-time offset, a transfer time, and a dwell time
-/// are all `Duration`; an arrival time and a departure time are
-/// both `SecondOfDay`.
-///
-/// Constructed via [`Duration::ZERO`], [`Duration::from_secs`], or
-/// the public-field constructor `Duration(n)`. Arithmetic is
-/// saturating: `Duration::MAX + Duration` stays at `Duration::MAX`.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
-pub struct Duration(pub u32);
-
-impl Duration {
-    /// Zero seconds.
-    pub const ZERO: Duration = Duration(0);
-    /// Maximum representable duration. Used internally as the
-    /// "no walk recorded" sentinel for some algorithm paths.
-    pub const MAX: Duration = Duration(u32::MAX);
-
-    /// Construct from a raw seconds count.
-    pub const fn from_secs(s: u32) -> Self {
-        Duration(s)
-    }
-
-    /// The underlying `u32` — seconds.
-    pub const fn as_secs(self) -> u32 {
-        self.0
-    }
-}
-
-impl From<u32> for Duration {
-    fn from(s: u32) -> Self {
-        Duration(s)
-    }
-}
-
-impl fmt::Display for Duration {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl std::ops::Add<Duration> for Duration {
-    type Output = Duration;
-    fn add(self, o: Duration) -> Duration {
-        Duration(self.0.saturating_add(o.0))
-    }
-}
-
-/// User-facing transfer cap. The algorithm explores rounds 0
-/// through `transfers` inclusive, so `Transfers(10)` lets a journey
-/// involve up to 10 boardings. `u8` is plenty — practical journey
-/// queries cap at single digits.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
-pub struct Transfers(pub u8);
-
-impl Transfers {
-    /// No transfers allowed (only direct journeys).
-    pub const ZERO: Transfers = Transfers(0);
-    /// Maximum representable transfer cap (255).
-    pub const MAX: Transfers = Transfers(u8::MAX);
-
-    /// Construct from a raw `u8`.
-    pub const fn new(n: u8) -> Self {
-        Transfers(n)
-    }
-
-    /// The underlying `u8`.
-    pub const fn get(self) -> u8 {
-        self.0
-    }
-}
-
-impl From<u8> for Transfers {
-    fn from(n: u8) -> Self {
-        Transfers(n)
-    }
-}
-
-impl fmt::Display for Transfers {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-/// Dense index of a stop within a [`Timetable`]. Indices are in `0..tt.n_stops()`.
-///
-/// Constructed by adapters at timetable-construction time. Display formats as
-/// the bare `u32`; round-trip via [`StopIdx::get`] / [`From<u32>`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct StopIdx(u32);
-
-impl StopIdx {
-    /// Construct from a raw `u32`. The caller is responsible for the value
-    /// being a valid index for the timetable in question.
-    pub const fn new(n: u32) -> Self {
-        Self(n)
-    }
-    /// The underlying `u32`.
-    pub const fn get(self) -> u32 {
-        self.0
-    }
-    #[inline]
-    pub(crate) fn idx(self) -> usize {
-        self.0 as usize
-    }
-}
-
-impl fmt::Display for StopIdx {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<u32> for StopIdx {
-    fn from(n: u32) -> Self {
-        Self(n)
-    }
-}
-impl From<StopIdx> for u32 {
-    fn from(s: StopIdx) -> Self {
-        s.0
-    }
-}
-
-/// Dense index of a route within a [`Timetable`]. Indices are in `0..tt.n_routes()`.
-///
-/// In the GTFS adapter, a single GTFS `route_id` may map to multiple
-/// `RouteIdx`s — one per equivalence class of trips with identical stop
-/// sequences and pairwise non-overtaking schedules. See
-/// [`gtfs::GtfsTimetable`] for the splitting rules and lookup APIs.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct RouteIdx(u32);
-
-impl RouteIdx {
-    /// Construct from a raw `u32`. The caller is responsible for the value
-    /// being a valid index for the timetable in question.
-    pub const fn new(n: u32) -> Self {
-        Self(n)
-    }
-    /// The underlying `u32`.
-    pub const fn get(self) -> u32 {
-        self.0
-    }
-    #[inline]
-    pub(crate) fn idx(self) -> usize {
-        self.0 as usize
-    }
-}
-
-impl fmt::Display for RouteIdx {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<u32> for RouteIdx {
-    fn from(n: u32) -> Self {
-        Self(n)
-    }
-}
-impl From<RouteIdx> for u32 {
-    fn from(r: RouteIdx) -> Self {
-        r.0
-    }
-}
-
-/// Dense index of a trip within a [`Timetable`]. Indices are in `0..n_trips`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct TripIdx(u32);
-
-impl TripIdx {
-    /// Construct from a raw `u32`. The caller is responsible for the value
-    /// being a valid index for the timetable in question.
-    pub const fn new(n: u32) -> Self {
-        Self(n)
-    }
-    /// The underlying `u32`.
-    pub const fn get(self) -> u32 {
-        self.0
-    }
-    #[inline]
-    pub(crate) fn idx(self) -> usize {
-        self.0 as usize
-    }
-}
-
-impl fmt::Display for TripIdx {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<u32> for TripIdx {
-    fn from(n: u32) -> Self {
-        Self(n)
-    }
-}
-impl From<TripIdx> for u32 {
-    fn from(t: TripIdx) -> Self {
-        t.0
-    }
-}
-
-/// A bag of `(stop, walk-time-offset)` pairs used as the origin or
-/// target of a query.
-///
-/// Constructed via the [`IntoEndpoints`] trait, which is implemented
-/// for the natural input shapes — a single stop, a slice of stops,
-/// a slice of `(stop, duration)` pairs, the owning `Vec` forms.
-/// Most callers don't construct `Endpoints` directly; they pass
-/// whatever they have to a query method that takes `impl IntoEndpoints`.
-///
-/// Stored inline up to 50 entries — covers the typical
-/// parent-station case (the largest real example we've seen is
-/// Berlin Hauptbahnhof at ~301 child platforms, which spills to the
-/// heap; everything else stays inline).
-#[derive(Debug, Clone, Default)]
-pub struct Endpoints {
-    stops: SmallVec<[(StopIdx, Duration); 50]>,
-}
-
-impl Endpoints {
-    /// Construct empty.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Number of `(stop, walk)` pairs.
-    pub fn len(&self) -> usize {
-        self.stops.len()
-    }
-
-    /// True iff this endpoint set has no stops.
-    pub fn is_empty(&self) -> bool {
-        self.stops.is_empty()
-    }
-
-    /// Borrow the underlying `(stop, walk)` slice.
-    pub fn as_slice(&self) -> &[(StopIdx, Duration)] {
-        &self.stops
-    }
-
-    /// Push a single `(stop, walk)` pair.
-    pub fn push(&mut self, stop: StopIdx, walk: Duration) {
-        self.stops.push((stop, walk));
-    }
-}
-
-/// Convert any of the natural query-endpoint inputs into an
-/// [`Endpoints`] value the algorithm can consume.
-///
-/// Implemented for:
-/// - `StopIdx` — a single stop, walk-time = 0.
-/// - `(StopIdx, Duration)` — a single stop with explicit walk-time.
-/// - `&[StopIdx]`, `&[StopIdx; N]` — multiple stops, all walk = 0.
-/// - `&[(StopIdx, Duration)]`, `&[(StopIdx, Duration); N]` — multiple
-///   stops with explicit walks.
-/// - `Vec<StopIdx>`, `Vec<(StopIdx, Duration)>` — owned variants.
-/// - `&Endpoints`, `Endpoints` — already-built value, identity.
-///
-/// For anything else (an iterator, your own collection type),
-/// collect into one of the above first or build an [`Endpoints`]
-/// directly via [`Endpoints::push`].
-pub trait IntoEndpoints {
-    /// Convert to an [`Endpoints`] value.
-    fn into_endpoints(self) -> Endpoints;
-}
-
-impl IntoEndpoints for StopIdx {
-    fn into_endpoints(self) -> Endpoints {
-        let mut e = Endpoints::new();
-        e.push(self, Duration::ZERO);
-        e
-    }
-}
-
-impl IntoEndpoints for (StopIdx, Duration) {
-    fn into_endpoints(self) -> Endpoints {
-        let mut e = Endpoints::new();
-        e.push(self.0, self.1);
-        e
-    }
-}
-
-impl IntoEndpoints for &[StopIdx] {
-    fn into_endpoints(self) -> Endpoints {
-        let mut e = Endpoints::new();
-        for &s in self {
-            e.push(s, Duration::ZERO);
-        }
-        e
-    }
-}
-
-impl<const N: usize> IntoEndpoints for &[StopIdx; N] {
-    fn into_endpoints(self) -> Endpoints {
-        (self.as_slice()).into_endpoints()
-    }
-}
-
-impl IntoEndpoints for &[(StopIdx, Duration)] {
-    fn into_endpoints(self) -> Endpoints {
-        let mut e = Endpoints::new();
-        for &p in self {
-            e.push(p.0, p.1);
-        }
-        e
-    }
-}
-
-impl<const N: usize> IntoEndpoints for &[(StopIdx, Duration); N] {
-    fn into_endpoints(self) -> Endpoints {
-        (self.as_slice()).into_endpoints()
-    }
-}
-
-impl IntoEndpoints for Vec<StopIdx> {
-    fn into_endpoints(self) -> Endpoints {
-        (self.as_slice()).into_endpoints()
-    }
-}
-
-impl IntoEndpoints for &Vec<StopIdx> {
-    fn into_endpoints(self) -> Endpoints {
-        (self.as_slice()).into_endpoints()
-    }
-}
-
-impl IntoEndpoints for Vec<(StopIdx, Duration)> {
-    fn into_endpoints(self) -> Endpoints {
-        let mut e = Endpoints::new();
-        for p in self {
-            e.push(p.0, p.1);
-        }
-        e
-    }
-}
-
-impl IntoEndpoints for &Vec<(StopIdx, Duration)> {
-    fn into_endpoints(self) -> Endpoints {
-        (self.as_slice()).into_endpoints()
-    }
-}
-
-impl IntoEndpoints for &Endpoints {
-    fn into_endpoints(self) -> Endpoints {
-        self.clone()
-    }
-}
-
-impl IntoEndpoints for Endpoints {
-    fn into_endpoints(self) -> Endpoints {
-        self
-    }
-}
-
-/// A label attached to a `(round, stop)` cell during the RAPTOR scan.
-///
-/// **Most users can ignore this trait.** [`Timetable::query`] uses
-/// [`ArrivalTime`] (single-criterion: minimise arrival time, fewest
-/// transfers), which is what the original RAPTOR paper describes and
-/// what almost every routing application wants.
-///
-/// The trait exists so the algorithm can be reused for *multi-criterion*
-/// routing — minimising arrival time *and* something else at the same
-/// time, returning a Pareto front of trade-offs. Reach for it when a
-/// single "best" answer is the wrong shape: e.g. an accessibility-aware
-/// query that should also report the route with less walking, even if
-/// it arrives slightly later. The bundled [`labels::ArrivalAndWalk`]
-/// is one such impl; see also [`Timetable::query_with_label`] for the
-/// builder entry point.
-///
-/// The algorithm maintains a Pareto front (a *bag* of mutually
-/// non-dominated labels) per `(round, stop)`, so multi-criterion impls
-/// produce real Pareto fronts at the targets rather than a single
-/// tiebroken label. Single-criterion `ArrivalTime` bags stay size 1,
-/// with no behaviour change versus a non-bag implementation.
-pub trait Label: Copy + std::fmt::Debug {
-    /// The "unreached" sentinel. The algorithm initialises every
-    /// `(round, stop)` cell to this value before seeding origins.
-    const UNREACHED: Self;
-
-    /// Initial label at an origin stop, given the user's departure time.
-    fn from_departure(at: SecondOfDay) -> Self;
-
-    /// New label produced by alighting from a trip at this stop with
-    /// the given arrival time. `self` is the label at the boarding
-    /// stop. For multi-criterion impls, components like accumulated
-    /// walking time inherit from `self`.
-    fn extend_by_trip(self, arrival: SecondOfDay) -> Self;
-
-    /// New label after walking a footpath of duration `walk_time`.
-    fn extend_by_footpath(self, walk_time: Duration) -> Self;
-
-    /// `self` weakly dominates `other` (every criterion of `self` is
-    /// at most the corresponding criterion of `other`). The default
-    /// implementation uses [`Label::arrival`], which is correct for
-    /// single-criterion impls.
-    fn dominates(&self, other: &Self) -> bool {
-        self.arrival() <= other.arrival()
-    }
-
-    /// Effective arrival time at the labelled stop. Used by the
-    /// algorithm for target-threshold comparisons and by [`Journey`]
-    /// output. Always returns [`SecondOfDay::MAX`] for [`Label::UNREACHED`].
-    fn arrival(&self) -> SecondOfDay;
-}
-
-/// Single-criterion label = arrival time at a stop. Default `L`
-/// throughout the algorithm. Constructing from a `SecondOfDay` is direct;
-/// extracting back is `arrival()`.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct ArrivalTime(pub SecondOfDay);
-
-impl Label for ArrivalTime {
-    const UNREACHED: Self = ArrivalTime(SecondOfDay::MAX);
-
-    #[inline]
-    fn from_departure(at: SecondOfDay) -> Self {
-        ArrivalTime(at)
-    }
-
-    #[inline]
-    fn extend_by_trip(self, arrival: SecondOfDay) -> Self {
-        ArrivalTime(arrival)
-    }
-
-    #[inline]
-    fn extend_by_footpath(self, walk_time: Duration) -> Self {
-        ArrivalTime(self.0 + walk_time)
-    }
-
-    #[inline]
-    fn arrival(&self) -> SecondOfDay {
-        self.0
-    }
-}
 
 /// A journey found by the RAPTOR algorithm.
 ///
